Refactor; reduce repetition;; small changes to text files
